ArrayQueue based file logging.
New method: We use Crossbeam ArrayQueue as a multithreaded buffer.
We can write log messages into this buffer from any thread without
locking. When the queue is close to full we get a lock to the underlying
file, drain the queue, and write to the file. During this time other
threads are still able to commit messages to the buffer.

Old method. When a log had to be written the thread would get a lock to
the Write, build the log entry, write to the Write, release the lock. If
multiple threads have to write at the same time it would cause lock
contention.